upgrade Go to 1.26.1 and all dependencies
Upgrade Go from 1.22.5 to 1.26.1. Upgrade urfave/cli/v3 from
v3.0.0-alpha9 to v3.8.0 stable (int64->int for IntFlag, BeforeFunc
now returns context). Upgrade echo v4.11.4->v4.15.1, testify
v1.9.0->v1.11.1, x/net v0.27.0->v0.52.0, and all transitive deps.

Replace deprecated golang.org/x/net/context with stdlib context.
Add tidy and update-deps recipes to Justfile.

+++ cli/cli_flags_test.go
@@ -33,11 +33,11 @@ func TestWorkers(t *testing.T) {
 func TestRetries(t *testing.T) {
 	results, _ := ParseGandaArgs([]string{"ganda"})
 	assert.NotNil(t, results)
-	assert.Equal(t, int64(0), results.GetContext().Retries)
+	assert.Equal(t, 0, results.GetContext().Retries)
 
 	separateResults, _ := ParseGandaArgs([]string{"ganda", "--retry", "5"})
 	assert.NotNil(t, separateResults)
-	assert.Equal(t, int64(5), separateResults.GetContext().Retries)
+	assert.Equal(t, 5, separateResults.GetContext().Retries)
 }
 
 func TestInvalidWorkers(t *testing.T) {
